test: move iOS emulation test to detectBrowser.test.js
